+++ src/pm-conversations/pmConversationsSelectors.js
@@ -3,7 +3,7 @@ import { createSelector } from 'reselect';
 
 import type { Message, PmConversationData } from '../types';
 import { getPrivateMessages } from '../baseSelectors';
-import { getOwnEmail } from '../account/accountSelectors';
+import { getOwnEmail } from '../account/accountsSelectors';
 import { getUnreadByPms, getUnreadByHuddles } from '../unread/unreadSelectors';
 import { normalizeRecipientsSansMe, getRecipientsIds } from '../utils/recipient';
