+++ src/unread/unreadSelectors.js
@@ -13,7 +13,7 @@ import {
   getUnreadMentions,
 } from '../directSelectors';
 import { getPrivateMessages } from '../baseSelectors';
-import { getOwnEmail } from '../account/accountSelectors';
+import { getOwnEmail } from '../account/accountsSelectors';
 import { getSubscriptionsById } from '../subscriptions/subscriptionSelectors';
 import { countUnread } from '../utils/unread';
 import { isTopicMuted } from '../utils/message';
